Removed Update Button

+++ tutorial/flaskr/schema.sql
@@ -10,14 +10,6 @@ CREATE TABLE bankacc (
   last_name VARCHAR (50) NOT NULL,
   username TEXT UNIQUE NOT NULL,
   password TEXT NOT NULL,
-  balance DECIMAL(15,2) NOT NULL
+  balance FLOAT(15,2) NOT NULL
 );
 
--- CREATE TABLE post (
---   id INTEGER PRIMARY KEY AUTOINCREMENT,
---   author_id INTEGER NOT NULL,
---   created TIMESTAMP NOT NULL DEFAULT CURRENT_TIMESTAMP,
---   title TEXT NOT NULL,
---   body TEXT NOT NULL,
---   FOREIGN KEY (author_id) REFERENCES user (id)
--- );
